pyproject.toml
[tool.poetry]
name = "DiscordTimeFinder"
version = "0.1.0"
description = "Find a suitable time together"
authors = ["ncrqnt <nico@ncrqnt.ch>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
orjson = "^3.6.7"
naff = "^1.0.0"
pytz = "^2022.1"

[tool.poetry.dev-dependencies]
flake8 = "*"
autopep8 = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
